requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bmtrain=0.1.8=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
datasets=1.17.0=pypi_0
deepspeed=0.6.0=pypi_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
transformers=4.18.0=pypi_0
